Mosquitto ssl (#6)
This makes it possible to run mosquitto with SSL. It preserves all of the SSL stuff that nginx-proxy uses, but generalizes it to make it able to create other certs more easily.

It also adds scripts to build the files needed for mosquitto password authentication. The dynamic security plugin for mosquitto looks cool, but it's overkill for what I need here.