Configurable domain (#16)
This adds support for configuring the pihome domain. Without these changes, the whole pihome setup is hard-coded to pihome.run. These changes allow users to specify the domain with some env vars.

this commit is my solution to a DNS snag where the containers were unable to use the DNS server running on the host. I'm not positive it's the correct way to resolve it 9b9fe1e


* specify the hostname of the DNS server when binding the port

I was seeing strange DNS behavior here. I was able to resolve DNS on the
host through `nslookup <some hostname> 10.0.0.10 (the location of the
host running pihole)`, and I was able to hit the DNS port from within
the containers (`nc -vz 10.0.0.10 53` succeeded inside the containers),
but I couldnt do DNS lookups against 10.0.0.10 from inside the
containers. I saw error messages that looked like:

```
>  nslookup google.com 10.0.0.10
;; reply from unexpected source: 172.22.0.1#53, expected 10.0.0.10#53

;; reply from unexpected source: 172.22.0.1#53, expected 10.0.0.10#5
```

I'm not totally sure why the reply from 127.22.0.1 (the gateway host of
the docker network that this whole thing is running on) was unexpected,
but it was causing problems. binding the host's IP addres resolved these
issues though

see https://forums.docker.com/t/dns-issues-with-local-resolver-and-containers-on-the-same-host/102319/4